fix: lazy load entirety of learning path edit page
