+++ src/containers/MyNdla/Learningpath/utils.ts
@@ -8,9 +8,7 @@
 
 import config from "../../../config";
 import { GQLMyNdlaLearningpathStepFragment } from "../../../graphqlTypes";
-import { deserializeToRichText, serializeFromRichText } from "../../../components/RichTextEditor/richTextSerialization";
 import { FormValues } from "./types";
-import { unreachable } from "../../../util/guards";
 
 export const sharedLearningpathLink = (id: number) => `${config.ndlaFrontendDomain}/learningpaths/${id}`;
 
@@ -26,70 +24,6 @@ export const getFormTypeFromStep = (step?: GQLMyNdlaLearningpathStepFragment): F
   if (step?.resource || step.embedUrl?.url.includes("resource")) return "resource";
   if (step?.embedUrl?.embedType === "external") return "external";
   return "text";
-};
-
-export const toFormValues = <T extends FormValues["type"]>(
-  type: T,
-  step?: GQLMyNdlaLearningpathStepFragment,
-): FormValues => {
-  switch (type) {
-    case "text":
-      return {
-        type: "text",
-        title: step?.title ?? "",
-        introduction: step?.introduction ?? "",
-        description: deserializeToRichText(step?.description ?? ""),
-      };
-    case "external":
-      return {
-        type: type,
-        title: step?.title ?? "",
-        introduction: step?.introduction ?? "",
-        url: step?.embedUrl?.url ?? "",
-        shareable: !!step?.embedUrl?.url,
-      };
-    case "resource":
-    case "folder":
-      return {
-        type: type,
-        title: step?.title ?? "",
-        embedUrl: step?.embedUrl?.url ?? "",
-      };
-    default:
-      return unreachable(type);
-  }
-};
-
-export const formValuesToGQLInput = (values: FormValues) => {
-  if (values.type === "text") {
-    return {
-      type: "TEXT",
-      title: values.title,
-      introduction: values.introduction,
-      description: serializeFromRichText(values.description),
-    };
-  }
-
-  if (values.type === "external") {
-    return {
-      type: "TEXT",
-      title: values.title,
-      introduction: values.introduction,
-      embedUrl: {
-        url: values.url,
-        embedType: "external",
-      },
-    };
-  }
-
-  return {
-    type: "TEXT",
-    title: values.title,
-    embedUrl: {
-      url: values.embedUrl,
-      embedType: "iframe",
-    },
-  };
 };
 
 export const learningpathListItemId = (id: number) => `learningpath-${id}`;
